task/requirements.txt
aiohttp==3.11.11
aiomysql==0.2.0
PyYAML==6.0.2
PyYAML==6.0.2
